+++ test/libsolidity/semanticTests/empty_contract.sol
@@ -1,0 +1,4 @@
+contract test { }
+// ----
+// i_am_not_there()
+// REVERT